requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pydantic==1.10.12
torch==2.1.0 
torchvision==0.16.0
Pillow==9.5.0
matplotlib==3.7.1
python-multipart==0.0.6
